Trac Ticket #44.
This submission completes the correction of bogus error messages emitted by the ConsoleTestCPP program, especially the TestN component thereof.  In this submission, we deal with several issues related to the comparison of datum definitions in CS-MAP and EPSG.  In particular, we correct defects with regard to selection of the proper EPSG variant to which the CS-MAP definition is compared to.  We have not, as yuet, dealt with the issue of comparing Geodetic transformations.


git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1971 4e78687f-474d-0410-85f9-8d5e500ac6b2

+++ CsMapConsoleUtilities/Source/csAddEpsgData.cpp
@@ -168,7 +168,7 @@ bool csAddEpsgCodes (TcsDefFile& dictionaryAsc)
 
 	ok = true;
 	size_t defCount = dictionaryAsc.GetDefinitionCount () - 1;
-	for (size_t index = 0;ok && index < defCount;index += 1)
+	for (unsigned index = 0;ok && index < defCount;index += 1)
 	{
 		ascDefPtr = &(*dictionaryAsc [index]);
 		if (ascDefPtr == 0)
@@ -287,7 +287,7 @@ bool csAddSridCodes (TcsDefFile& dictionaryAsc)
 
 	ok = true;
 	size_t defCount = dictionaryAsc.GetDefinitionCount () - 1;
-	for (size_t index = 0;ok && index < defCount;index += 1)
+	for (unsigned index = 0;ok && index < defCount;index += 1)
 	{
 		ascDefPtr = &(*dictionaryAsc [index]);
 		if (ascDefPtr == 0)
@@ -395,7 +395,7 @@ bool csAddEpsgQuadrant (TcsDefFile& dictionaryAsc,const TcsEpsgDataSetV6& epsgDa
 	ok = true;
 	dictType = dictionaryAsc.GetDictType ();
 	size_t defCount = dictionaryAsc.GetDefinitionCount () - 1;
-	for (size_t index = 0;ok && index < defCount;index += 1)
+	for (unsigned index = 0;ok && index < defCount;index += 1)
 	{
 		ascDefPtr = &(*dictionaryAsc [index]);
 		if (ascDefPtr == 0)
